requirements.txt
Django==5.0.3
gunicorn
dj-database-url
psycopg2-binary
python-dotenv
whitenoise
cloudinary
django-cloudinary-storage
django-cors-headers
djangorestframework
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
cryptography
python-dotenv

